namespace widget.rb
