+++ test/unit/widget_test.rb
@@ -2,7 +2,7 @@ require "test_helper"
 
 class WidgetTest < Test::Unit::TestCase
   def test_widget
-    widget = Widget.find(:test)
+    widget = Chameleon::Widget.find(:test)
     assert_not_nil widget
     assert_equal :test, widget.name
     assert_equal "c6fdeae2af2c327cac35abc2ccacd8619af56821", widget.key
